rename createDispatcher.js to dispatcher.js
